Problem with update_player_score function
